Fix harmless -Wsign-compare gcc warnings in wxQt wxStatusBar code
QList::count() returns int, so cast it to size_t explicitly before
comparing with wxArray::GetCount() which returns size_t.

+++ src/qt/statusbar.cpp
@@ -65,7 +65,7 @@ bool wxStatusBar::GetFieldRect(int i, wxRect& rect) const
     wxCHECK_MSG( (i >= 0) && ((size_t)i < m_panes.GetCount()), false,
                  "invalid statusbar field index" );
 
-    if ( m_qtPanes->count() != m_panes.GetCount() )
+    if ( static_cast<size_t>(m_qtPanes->count()) != m_panes.GetCount() )
         const_cast<wxStatusBar*>(this)->UpdateFields();
     
     rect = wxQtConvertRect((*m_qtPanes)[i]->geometry());
@@ -89,7 +89,7 @@ int wxStatusBar::GetBorderY() const
 
 void wxStatusBar::DoUpdateStatusText(int number)
 {
-    if ( m_qtPanes->count() != m_panes.GetCount() )
+    if ( static_cast<size_t>(m_qtPanes->count()) != m_panes.GetCount() )
         UpdateFields();
 
     (*m_qtPanes)[number]->setText( wxQtConvertString( m_panes[number].GetText() ) );
